requirements.txt
GitPython
requests
pandas
pyyaml
rich
python-dotenv
